+++ src/routes/routes.tsx
@@ -95,7 +95,6 @@ const appRoutes: (Route | NavLinkRoute)[] = [
   },
   {
     path: `/${path('content_page')}/:slug`,
-    title: rt('content_page'),
     component: () => <ContentPage />,
     exact: true,
     isPublic: true,
@@ -109,14 +108,12 @@ const appRoutes: (Route | NavLinkRoute)[] = [
   },
   {
     path: `/${path('tests')}/:slug`,
-    title: rt('tests'),
     component: () => <Test />,
     exact: true,
     isPublic: true,
   },
   {
     path: `/${path('tests')}/:slug/${path('outcome')}`,
-    title: rt('outcome'),
     component: () => <TestOutcomes />,
     exact: true,
     isPublic: true,